+++ opencv/src/main/java/org/bytedeco/javacpp/presets/opencv_cudaoptflow.java
@@ -45,7 +45,8 @@ import org.bytedeco.javacpp.tools.InfoMapper;
         ),
         @Platform(
             value = "windows",
-            link = "opencv_cudaoptflow341",
+            link = "opencv_cudaoptflow342",
+            preload = "opencv_cudalegacy342",
             extension = "-gpu"
         )
     },