+++ opencv/src/main/java/org/bytedeco/javacpp/presets/opencv_plot.java
@@ -36,7 +36,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
             link = "opencv_plot@.3.4"
         ),
         @Platform(value = "ios", preload = "libopencv_plot"),
-        @Platform(value = "windows", link = "opencv_plot341")
+        @Platform(value = "windows", link = "opencv_plot342")
     },
     target = "org.bytedeco.javacpp.opencv_plot"
 )