+++ opencv/src/main/java/org/bytedeco/javacpp/presets/opencv_saliency.java
@@ -41,7 +41,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
             link = "opencv_saliency@.3.4"
         ),
         @Platform(value = "ios", preload = "libopencv_saliency"),
-        @Platform(value = "windows", link = "opencv_saliency341")
+        @Platform(value = "windows", link = "opencv_saliency342")
     },
     target = "org.bytedeco.javacpp.opencv_saliency"
 )