+++ opencv/src/main/java/org/bytedeco/javacpp/presets/opencv_superres.java
@@ -37,10 +37,10 @@ import org.bytedeco.javacpp.tools.InfoMapper;
               preload = {"opencv_cuda@.3.4", "opencv_cudacodec@.3.4", "opencv_cudaarithm@.3.4", "opencv_cudafilters@.3.4",
                          "opencv_cudaimgproc@.3.4", "opencv_cudafeatures2d@.3.4", "opencv_cudalegacy@.3.4",
                          "opencv_cudaoptflow@.3.4", "opencv_cudawarping@.3.4"}),
-    @Platform(value = "windows", link = "opencv_superres341",
-              preload = {"opencv_cuda341", "opencv_cudacodec341", "opencv_cudaarithm341", "opencv_cudafilters341",
-                         "opencv_cudaimgproc341", "opencv_cudafeatures2d341", "opencv_cudalegacy341",
-                         "opencv_cudaoptflow341", "opencv_cudawarping341"})},
+    @Platform(value = "windows", link = "opencv_superres342",
+              preload = {"opencv_cuda342", "opencv_cudacodec342", "opencv_cudaarithm342", "opencv_cudafilters342",
+                         "opencv_cudaimgproc342", "opencv_cudafeatures2d342", "opencv_cudalegacy342",
+                         "opencv_cudaoptflow342", "opencv_cudawarping342"})},
         target = "org.bytedeco.javacpp.opencv_superres")
 public class opencv_superres implements InfoMapper {
     public void map(InfoMap infoMap) {